Moved service comparator tests for access to service deployer
