+++ core/framework/comparator_test.go
@@ -803,131 +803,3 @@ func Test_CreateRouteComparator(t *testing.T) {
 		})
 	}
 }
-
-func Test_CreateServiceComparator(t *testing.T) {
-	type args struct {
-		deployed  resource.KubernetesResource
-		requested resource.KubernetesResource
-	}
-	familyPolicy := v1.IPFamilyPolicyRequireDualStack
-	tests := []struct {
-		name  string
-		args  args
-		want  reflect.Type
-		want1 bool
-	}{
-		{
-			"Equals",
-			args{
-				deployed: &v1.Service{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"app":     "test",
-							"service": "test",
-						},
-					},
-				},
-				requested: &v1.Service{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"app":     "test",
-							"service": "test",
-						},
-					},
-				},
-			},
-			reflect.TypeOf(v1.Service{}),
-			true,
-		},
-		{
-			"NotEquals",
-			args{
-				deployed: &v1.Service{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"app":     "test",
-							"service": "test",
-						},
-					},
-				},
-				requested: &v1.Service{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"app":     "test",
-							"service": "test1",
-						},
-					},
-				},
-			},
-			reflect.TypeOf(v1.Service{}),
-			false,
-		},
-		{
-			"DifferentClusterIPs",
-			args{
-				deployed: &v1.Service{
-					Spec: v1.ServiceSpec{
-						ClusterIPs: []string{"10.217.5.142"},
-					},
-				},
-				requested: &v1.Service{},
-			},
-			reflect.TypeOf(v1.Service{}),
-			true,
-		},
-		{
-			"DifferentClusterIP",
-			args{
-				deployed: &v1.Service{
-					Spec: v1.ServiceSpec{
-						ClusterIP: "10.217.5.142",
-					},
-				},
-				requested: &v1.Service{},
-			},
-			reflect.TypeOf(v1.Service{}),
-			true,
-		},
-		{
-			"DifferentIPFamilies",
-			args{
-				deployed: &v1.Service{
-					Spec: v1.ServiceSpec{
-						IPFamilies: []v1.IPFamily{"IPv4"},
-					},
-				},
-				requested: &v1.Service{},
-			},
-			reflect.TypeOf(v1.Service{}),
-			true,
-		},
-		{
-			"DifferentIPFamilyPolicy",
-			args{
-				deployed: &v1.Service{
-					Spec: v1.ServiceSpec{
-						IPFamilyPolicy: &familyPolicy,
-					},
-				},
-				requested: &v1.Service{},
-			},
-			reflect.TypeOf(v1.Service{}),
-			true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, got1 :=
-				NewComparatorBuilder().
-					WithType(tt.want).
-					WithCustomComparator(CreateServiceComparator()).
-					Build()
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("createServiceComparator() got = %v, want %v", got, tt.want)
-			}
-			if !reflect.DeepEqual(got1(tt.args.deployed, tt.args.requested), tt.want1) {
-				t.Errorf("createServiceComparator() got1 = %v, want %v", got1(tt.args.deployed, tt.args.requested), tt.want1)
-			}
-		})
-	}
-}
